backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
supabase>=2.4.0
python-dotenv>=1.0.1
werkzeug>=3.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.1.1
requests>=2.31.0
scalar-fastapi>=1.0.0